# public/docs/api-docs.yaml
openapi: 3.0.0
info:
  title: API de Gestion des Articles
  version: 1.0.0
  description: Documentation de l'API pour la gestion des articles, des utilisateurs, des clients , des dettes et leur paiement et de l'authentification
servers:
  - url: http://localhost:3000/wane/v1
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      bearer: 
  schemas:
...
    Article:
      type: object
      properties:
        id:
          type: integer
        libelle:
          type: string
        price:
          type: float
    StoreArticleRequest:
      type: object
      properties:
        libelle:
          type: string
    ArticleUpdateRequest:
      type: object
      properties:
        libelle:
          type: string
    UpdateStockRequest:
      type: object
      properties:
        updated_articles:
          type: array
          items:
            type: object
            properties:
              article_id:
                type: integer
              qte:
                type: integer
      required:
        - updated_articles
...
  responses:
    UnauthorizedError:
      description: Authentification échouée, accès non autorisé
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
paths:
...
  /articles:
    get:
      security:
        - BearerAuth: []
      summary: Afficher la liste des articles
      operationId: getArticles
      tags:
        - Articles
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Numéro de la page à afficher
        - in: query
          name: per_page
          schema:
            type: integer
            default: 15
          description: Nombre d'articles par page
        - in: query
          name: disponible
          schema:
            type: string
            enum: [oui, non]
          description: Filtrer les articles par disponibilité
        - in: query
          name: trashed
          schema:
            type: string
            enum: [oui, non]
          description: Filtrer les articles supprimés
      responses:
        '200':
          description: Liste des articles récupérée avec succès
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Article'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      security:
        - BearerAuth: []
      summary: Créer un nouvel article
      operationId: createArticle
      tags:
        - Articles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreArticleRequest'
      responses:
        '201':
          description: Article créé avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'
        '500':
          description: Erreur interne du serveur
  /articles/{id}:
    get:
      security:
        - BearerAuth: []
      summary: Afficher un article spécifique
      operationId: getArticle
      tags:
        - Articles
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID de l'article à afficher
      responses:
        '200':
          description: Article récupéré avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'
        '404':
          description: Article non trouvé
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    put:
      security:
        - BearerAuth: []
      summary: Mettre à jour un article spécifique
      operationId: updateArticle
      tags:
        - Articles
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID de l'article à mettre à jour
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleUpdateRequest'
      responses:
        '200':
          description: Article mis à jour avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'
        '500':
          description: Erreur interne du serveur
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    delete:
      security:
        - BearerAuth: []
      summary: Supprimer un article spécifique (SoftDelete)
      operationId: deleteArticle
      tags:
        - Articles
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: ID de l'article à supprimer
      responses:
        '204':
          description: Article supprimé avec succès
        '500':
          description: Erreur interne du serveur
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /articles/update-stock:
    post:
      security:
        - BearerAuth: []
      summary: Mettre à jour la quantité en stock de plusieurs articles
      operationId: updateStock
      tags:
        - Articles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStockRequest'
      responses:
        '200':
          description: Stock mis à jour avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated_articles:
                    type: array
                    items:
                      $ref: '#/components/schemas/Article'
                  errors:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        message:
                          type: string
        '500':
          description: Erreur interne du serveur
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...